Add files via upload
My first ever Django-based website. I wanted to do Django because I wanted to learn to to make a website that could have a python element - could've learnt Flask, but believed the benefits Django had was much more than Flask and that for me it would be much more useful to learn Django compared to Flask.